requirements.txt
pandas==2.2.2
requests==2.32.3
matplotlib==3.9.0
python-pptx==0.6.21
PyYAML==6.0.2